+++ webpack.main.config.js
@@ -3,7 +3,7 @@ module.exports = {
      * This is the main entry point for your application, it's the first file
      * that runs in the main process.
      */
-    entry: "./src/Main.ts",
+    entry: "./src/main/index.ts",
     // Put your normal webpack config below here
     module: {
         rules: require("./webpack.rules"),
